+++ NEWS.md
@@ -4,6 +4,23 @@ editor_options:
     wrap: 72
 ---
 
+# SIMplyBee version 0.4.2
+
+-   20??-??-??
+
+## Major changes
+-   TODO
+
+## New features
+-   TODO
+
+## Bug fixes
+-   TODO
+
+## Backgrond/package/etc. work
+-   We now removed setMisc() and getMisc() because we now use the new AlphaSimR
+structure of the misc slot that is much easier to use.
+
 # SIMplyBee version 0.4.1
 
 -   2024-09-19
@@ -26,7 +43,7 @@ which caused an error. We now read in the locations from a csv file.
     now c(0, 0) PR#500
 
 
-## New features ##
+## New features
 -   In setLocation(MultiColony) we can set one location (numeric) or
     multiple (list or data.frame) PR#500
 -   getLocation(MultiColony) got the collapse argument
@@ -56,8 +73,6 @@ which caused an error. We now read in the locations from a csv file.
 
 -   Bug fix - get\*Haplo() functions were returning diploid drones when
     input was a Pop-class
-
--   
 
 # SIMplyBee version 0.3.0